rev: 209
Everything is implemented so should be safe to use 🥂

Issues:
There may be a problem with getInteracting() when the actor is a Player. Needs testing.

+++ api/src/main/java/net/runelite/api/Scene.java
@@ -127,15 +127,17 @@ public interface Scene
 
 	/**
 	 * Get the underlay ids for the scene. The value stored is id + 1, with 0 for no underlay.
+	 *
 	 * @return
 	 */
-	byte[][][] getUnderlayIds();
+	short[][][] getUnderlayIds();
 
 	/**
-	 * Get the overlay ids for the scene. The value stored is id + 1, with 0 for no overlay.
-	 * @return
-	 */
-	byte[][][] getOverlayIds();
+   * Get the overlay ids for the scene. The value stored is id + 1, with 0 for no overlay.
+   *
+   * @return
+   */
+	short[][][] getOverlayIds();
 
 	/**
 	 * Get the shapes of the tiles for the scene.
